[ROSAPPS] Move imagesoft to applications and add it to build. Convert lang files to UTF-8. Fix casts and x64 warnings. Patch by Hermes Belusca. See bug 7185 for details.
svn path=/trunk/; revision=57201

+++ rosapps/applications/imagesoft/imgedwnd.c
@@ -350,8 +350,8 @@ ImageEditWndProc(HWND hwnd,
                      yMouse,
                      NULL);
 
-            xMouse = (short)LOWORD(lParam);
-            yMouse = (short)HIWORD(lParam);
+            xMouse = LOWORD(lParam);
+            yMouse = HIWORD(lParam);
 
             LineTo(hDC,
                    xMouse,
